requirements.txt
asgiref==3.9.1
boto3==1.39.17
botocore==1.39.17
certifi==2025.7.14
cffi==1.17.1
cryptography==45.0.5
django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-knox @ git+https://github.com/James1345/django-rest-knox.git@8447cd00358b1a3f76862278711abd2cdcb8837a#egg=django_rest_knox
django-resumable @ git+https://github.com/rphlo/django-resumable@b941d3406e3fcf73b3b4ffefc9572e7b55decb30
djangorestframework==3.16.0
gunicorn==23.0.0
jmespath==1.0.1
mutagen==1.47.0
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pycparser==2.22
python-dateutil==2.9.0.post0
python-magic==0.4.27
raven==6.10.0
s3transfer==0.13.1
sentry-sdk==2.34.1
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.14.1
urllib3==2.5.0
